requirements.txt
flask
flask-cors
pandas
openpyxl
supabase
python-dotenv
numpy
flask-limiter
